requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pymongo[srv]==4.10.1
sentence-transformers==2.7.0
lightgbm==4.5.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
